#484 - Rollup doesn't like the way the w3 js files are writtens so stick them all in one file
