+++ src/js/charts/cannabis-local-ordinances/src/index.js
@@ -1,4 +1,4 @@
 import "./js/county-map/index.js";
 import "./js/table-data/index.js";
-// import "./js/combo-box/index.js";
+import "./js/places/dist/index.js";
