+++ src/js/charts/cannabis-local-ordinances/src/js/places/README.md
@@ -2,14 +2,12 @@ Notes:
 
 - update src/js/charts/cannabis-local-ordinances/adr/03-autocomplete.md
 - credit w3.com
+- dynamically generate place list
+- wire selects to map/table actions
 
 Operate:
 
 ```diff
-# src/js/charts/cannabis-local-ordinances/src/index.js
-
-+ import "./js/places/dist/index.js";
-
 
 # src/templates/wordpress/pages/where-cannabis-business-is-legal-in-california.html
