+++ src/js/charts/cannabis-local-ordinances/src/js/places/rollup.config.js
@@ -2,12 +2,7 @@ import css from "rollup-plugin-import-css";
 import multi from "@rollup/plugin-multi-entry";
 
 export default {
-  input: [
-    "src/index.js",
-    "src/w3-combobox/combobox-1.0-list.js",
-    "src/w3-combobox/listbox.js",
-    "src/w3-combobox/listboxOption.js",
-  ],
+  input: ["src/index.js", "src/w3-combobox/combobox-1.0-list.js"],
   output: { file: "dist/index.js", format: "esm" },
   plugins: [css(), multi()],
 };
